+++ .utils/pytest.sh
@@ -1,3 +1,3 @@
 #! /bin/bash
 
-pytest -vvx --capture=tee-sys --log-cli-level=INFO tests/models
+pytest -vvx --capture=tee-sys --log-cli-level=INFO tests/