+++ chess/templates/tournament.py
@@ -1,6 +1,8 @@
 #### NO MODELS IN TEMPLATES !!!!      ####
 from typing import List, Dict
 
+import logging
+
 
 class TournamentTemplate:
     """Template for tournament management."""
@@ -8,20 +10,23 @@ class TournamentTemplate:
     @classmethod
     def menu(cls) -> str:
         """Display tournament menu options."""
+
         print("\nTournament Menu")
         print("1. Create a tournament")
         print("2. Add a player")
         print("3. Launch tournament")
         print("4. Create a new round")
         print("5. Display rankings")
-        print("6. Return to the main menu")
+        print("6. Return to the main menu")  # "" to return to the main menu ?
 
         return input("Enter the number you want: ")
 
     @classmethod
     def create(cls) -> Dict:
         """Template for creating a tournament."""
+
         print("\nCreating a new tournament...")
+
         name = input("Enter the name of the tournament: ")
         start_date = input("Enter the start date of the tournament (YYYY-MM-DD): ")
         end_date = input("Enter the end date of the tournament (YYYY-MM-DD): ")
@@ -37,9 +42,23 @@ class TournamentTemplate:
         }
 
     @classmethod
+    def update(cls, tournament: Dict) -> Dict:
+        """Template for updating a tournament."""
+
+        print("\nUpdating the tournament... NOT IMPLETED YET !")
+
+    @classmethod
+    def delete(cls, tournament: Dict) -> Dict:
+        """Template for deleting a tournament."""
+
+        print("\nDeleting the tournament... NOT IMPLETED YET !")
+
+    @classmethod
     def add_player(cls) -> Dict:
         """Template for adding a player to the tournament."""
+
         print("\nAdding a new player...")
+
         firstname = input("Enter the first name of the player: ")
         lastname = input("Enter the last name of the player: ")
         birthdate = input("Enter the birthdate of the player (YYYY-MM-DD): ")
@@ -51,88 +70,65 @@ class TournamentTemplate:
         }
 
     @classmethod
-    def launch(cls) -> str:
+    def launch(cls) -> bool:
         """Template for launching the tournament."""
+
         print("\nLaunching the tournament...")
-        return input("Press Enter to confirm launching the tournament.")
+        choice = input(
+            "Press Enter to confirm launching the tournament / Any key to cancel."
+        )
+
+        # if choice:
+        #    return False
+
+        # return True
+
+        return False if choice else True  # ternary operator
 
     @classmethod
     def new_round(cls) -> None:
         """Template for creating a new round."""
-        print("\nCreating a new round...")
+
+        print("\Going to netx round")
+        choice = input("Press Enter to confirm / Any key to cancel.")
+
+        return False if choice else True  # ternary operator
+
+    @classmethod
+    def update_current_round(cls, match_list: List[List]) -> List[List]:
+        """Template for updating the current round."""
+
+        # logging.info(f"BEFORE: {match_list}")
+
+        print("\nUpdating the current round...")
+
+        for n_match, match in enumerate(match_list):
+            print(f"Match {n_match}:")
+
+            # logging.info(f"match: {match}")
+
+            for n_player, player in enumerate(match):
+
+                # logging.info(f"player: {player}")
+                print(f"Enter the score of the player id {player[0]}: ")
+                score = int(input())
+
+                # logging.info(f"???: {match_list[n_match][n_player][1]} ==> {score}")
+                # AWFUL PATCH !!! => to fix in the models drirecty
+                match_list[n_match][n_player] = list(match_list[n_match][n_player])
+                match_list[n_match][n_player][1] = score
+
+        # logging.info(f"AFTER: {match_list}")
+
+        return match_list
 
     @classmethod
     def display_rankings(cls, rankings: List[Dict]) -> None:
         """Template for displaying rankings."""
+
         print("\nRankings:")
+
         for rank, player in enumerate(rankings, start=1):
-            print(f"{rank}. {player['firstname']} {player['lastname']} - {player['score']} points")
-
-
-
-# class TournamentManagementSystem:
-
-#     def __init__(self, tournament_class):
-#         self.tournament_class = tournament_class
-#         self.tournament = None
-#         self.players = []
-
-#     def create_tournament(self):
-#         print("Creating a new tournament...")
-#         name = input("Enter the name of the tournament: ")
-#         while True:
-#             start_date = input("Enter the start date of the tournament (YYYY-MM-DD): ")
-#             try:
-#                 datetime.datetime.strptime(start_date, "%Y-%m-%d")
-#                 break
-#             except ValueError:
-#                 print(
-#                     "Invalid date format. Please enter the date in YYYY-MM-DD format."
-#                 )
-#         while True:
-#             end_date = input("Enter the end date of the tournament (YYYY-MM-DD): ")
-#             try:
-#                 datetime.datetime.strptime(end_date, "%Y-%m-%d")
-#                 break
-#             except ValueError:
-#                 print(
-#                     "Invalid date format. Please enter the date in YYYY-MM-DD format."
-#                 )
-#         description = input("Enter the description of the tournament: ")
-#         location = input("Enter the location of the tournament: ")
-
-#         tournament_id = secrets.token_hex(2)
-
-#         self.tournament = self.tournament_class(
-#             name,
-#             start_date,
-#             end_date,
-#             description,
-#             location,
-#             tournament_id=tournament_id,
-#         )
-#         self.tournament.create()
-
-#         # Automatically create rounds
-#         self.auto_create_rounds()
-
-#         # Launch tournament if enough players
-#         if len(self.players) >= self.tournament.N_PLAYERS:
-#             self.launch_tournament()
-
-#     def auto_create_rounds(self):
-#         print("Automatically creating rounds...")
-#         for round_number in range(1, self.tournament.N_ROUNDS + 1):
-#             matches = []
-#             self.tournament.add_round(round_number, matches)
-#         print("Rounds created successfully.")
-
-#     def launch_tournament(self):
-#         if self.tournament:
-#             try:
-#                 self.tournament.update_status("In Progress")
-#                 print("Tournament launched successfully.")
-#             except ValueError as e:
-#                 print(e)
-#         else:
-#             print("No tournament available to launch.")
+            print(
+                f"{rank}. {player['firstname']} {player['lastname']} - {player['score']} points"
+            )